# codeintel: Speed up document canonicalization by 100x~300x